dpdk: move common and util functions and structs
